Add files via upload
I improve upon the webscraping program written previously to gather the Team Stats from Game 3 of the 2018 Western Conference Finals. The goal of this notebook is to serve as the middle ground between the working notebook written previously and writing a function that takes url as input and a DataFrame of team stats as output.